PLAT-6785 Move performance reading results commits (6cf6d1bd - 6dd17abf)
Summary:
Refactored ColumnDefinitionPacket:
 - moved out logic
 - improved performance of reading huge result by removing redundant checks (of datatype, signed/unsigned, etc.) for every row decoding

Test Plan: https://app.circleci.com/pipelines/github/memsql/S2-JDBC-Connector/345/workflows/e418616b-2770-4807-affc-4436cc90ec3c

Reviewers: amakarovych-ua, pmishchenko-ua

Reviewed By: pmishchenko-ua

Subscribers: engineering-list

JIRA Issues: PLAT-6785

Differential Revision: https://grizzly.internal.memcompute.com/D65670

+++ src/main/java/com/singlestore/jdbc/client/impl/StandardClient.java
@@ -22,7 +22,7 @@ import com.singlestore.jdbc.client.socket.impl.CompressOutputStream;
 import com.singlestore.jdbc.client.socket.impl.PacketReader;
 import com.singlestore.jdbc.client.socket.impl.PacketWriter;
 import com.singlestore.jdbc.client.socket.impl.ReadAheadBufferedStream;
-import com.singlestore.jdbc.client.util.MutableInt;
+import com.singlestore.jdbc.client.util.MutableByte;
 import com.singlestore.jdbc.export.ExceptionFactory;
 import com.singlestore.jdbc.export.MaxAllowedPacketException;
 import com.singlestore.jdbc.export.Prepare;
@@ -66,8 +66,8 @@ public class StandardClient implements Client, AutoCloseable {
 
   protected final ExceptionFactory exceptionFactory;
   private Socket socket;
-  private final MutableInt sequence = new MutableInt();
-  private final MutableInt compressionSequence = new MutableInt();
+  private final MutableByte sequence = new MutableByte();
+  private final MutableByte compressionSequence = new MutableByte();
   private final ReentrantLock lock;
   private final Configuration conf;
   private final HostAddress hostAddress;
